requirements.txt
Pillow == 9.0.1
colored == 1.4.3
